Refactor AllocationBlock, move to separate components and add hooks
